change "mod" to "use"
This change undoes a bit of a nonobvious mistake I was making; by
coercing cargo to see the error module as a sibling, I inlined the
ZakocmpError type in every module into which it was pulled. This caused
the ultimate top-level user to see every module as having a unique
ZakocmpError type - i.e. config::errors::ZakocmpError,
snapshot::errors::ZakocmpError, and violations::errors::ZakocmpError.

The correct way to do this appears to be through "use crate::..."