snapshot: Adjust testing snapshot headers
With `simple-zakopane.sh` gone, the last references to it are in the
unit tests, which were written bearing the output of
`simple-zakopane.sh` in mind. The snapshot headers were never relevant
to the operation of the unit tests, though, so the mention of
`simple-zakopane.sh` does nothing in particular. (The same is true of
its removal, but I consider it visually distracting.)